Answer query about why symbol is in ELF
Add a text box and button to let the user query why a symbol is in
ELF. Index into symbols to obtain sym.

In a loop, use referencedObjFile to obtain obj file referencing sym,
and then index obj file into archiveObjfileSymRefs to obtain sym
causing that objfile to be linked in, adding syms to a list

Reverse list and print to show a list of objfile:sym pairs causing the
user provided sym to be linked in.